src/api/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.70.1
uvicorn[standard]==0.15.0
gunicorn==20.1.0
ruamel.yaml
# python-multipart
# python-jose[cryptography]
# passlib[bcrypt]
# python-graphql-client
# pysftp
